Extract `io_error_context` into `error` sub-module
