+++ lib/common/common/src/universal_io/io_uring/runtime.rs
@@ -325,23 +325,3 @@ impl<T> IoUringResponse<T> {
         }
     }
 }
-
-pub fn io_error_context(err: io::Error, context: impl Into<String>) -> io::Error {
-    io::Error::new(err.kind(), IoErrorContext::new(err, context))
-}
-
-#[derive(Debug, thiserror::Error)]
-#[error("{context}: {error}")]
-pub struct IoErrorContext {
-    context: String,
-    error: io::Error,
-}
-
-impl IoErrorContext {
-    fn new(error: io::Error, context: impl Into<String>) -> Self {
-        Self {
-            context: context.into(),
-            error,
-        }
-    }
-}
